UHF-5265: Prevent users from changing their email address, migrate old invalid localhost email addresses to hel.fi
