+++ helfi_tunnistamo.install
@@ -7,6 +7,8 @@
 
 declare(strict_types = 1);
 
+use Drupal\user\Entity\User;
+
 /**
  * Implements hook_install().
  */
@@ -34,5 +36,26 @@ function helfi_tunnistamo_update_9001() : void {
     Drupal::service('module_installer')->install([
       'helfi_api_base',
     ]);
+  }
+}
+
+/**
+ * Convert previously autogenerated {sub}@localhost mails to {sub}@hel.fi.
+ */
+function helfi_tunnistamo_update_9003() : void {
+  $query = \Drupal::database()
+    ->select('users_field_data', 'u')
+    ->fields('u', ['uid'])
+    ->condition('mail', "%@localhost", 'LIKE');
+
+  $users = $query->execute()->fetchAll();
+
+  foreach ($users as $user) {
+    if ((!$account = User::load($user->uid)) || !$account->getEmail()) {
+      continue;
+    }
+    [$sub] = explode('@', $account->getEmail());
+    $account->setEmail(helfi_tunnistamo_create_email(['sub' => $sub]))
+      ->save();
   }
 }
