+++ helfi_tunnistamo.module
@@ -7,6 +7,7 @@
 
 declare(strict_types = 1);
 
+use Drupal\Core\Form\FormStateInterface;
 use Drupal\helfi_tunnistamo\Plugin\OpenIDConnectClient\Tunnistamo;
 use Drupal\openid_connect\Entity\OpenIDConnectClientEntity;
 use Drupal\user\UserInterface;
@@ -26,7 +27,16 @@ function helfi_tunnistamo_openid_connect_post_authorize(UserInterface $account, 
 }
 
 /**
- * Generates an email address for given userinfo array.
+ * Implements hook_form_FORM_ID_alter().
+ */
+function helfi_tunnistamo_form_user_form_alter(&$form, FormStateInterface $form_state, $form_id) : void {
+  // Prevent users from changing the email.
+  $form['account']['mail']['#type'] = 'item';
+  $form['account']['mail']['#markup'] = $form['account']['mail']['#default_value'] ?? '';
+}
+
+/**
+ * Generates an email address for the given userinfo array.
  *
  * @param array $userinfo
  *   The userinfo.
@@ -35,7 +45,7 @@ function helfi_tunnistamo_openid_connect_post_authorize(UserInterface $account, 
  *   The autogenerated email.
  */
 function helfi_tunnistamo_create_email(array $userinfo) : string {
-  return $userinfo['sub'] . '@localhost';
+  return $userinfo['sub'] . '@hel.fi';
 }
 
 /**
@@ -51,8 +61,8 @@ function helfi_tunnistamo_openid_connect_userinfo_alter(
     return;
   }
 
-  // If client has 'ad_groups' scope and user email is missing, replace it with
-  // just some made-up address (issue with @edu.hel.fi users).
+  // If a client has 'ad_groups' scope and user email is missing, replace it
+  // with a made-up address (issue with @edu.hel.fi users).
   if (in_array('ad_groups', $plugin->getClientScopes()) && empty($userinfo['email'])) {
     $userinfo['email'] = helfi_tunnistamo_create_email($userinfo);
   }
